added permissions, tests are passing
